refactor: move InteractionCreateAction patch to own folder
